Create Bartek-checkpoint.ipynb
I have added the analysis of several regression models to the previous version of the file.

I have not plotted anything, I'm planning to do it in the next few days. I will also try testing some ensemble learning approaches.

I also did not do cross validation. It will probably be good to do CV in the final version of the project.